Use native-width types (usize/isize) for syscall arguments
Replace u64/i64 with usize/isize throughout the syscall layer so the
code works correctly on both 32-bit and 64-bit architectures without
unnecessary widening. On 64-bit, usize IS u64 (zero overhead). On
future 32-bit targets (armv7, i386), this avoids register pair
packing and 64-bit arithmetic for values that are inherently
pointer-width.

- fd, buf, len, size args → usize
- return values → isize
- File offsets (lseek) stay i64 (can exceed 4GB even on 32-bit)
- dispatch() signature: (Syscall, usize×5) → isize
- Arch entry points cast from register-width u64 to usize at boundary

+++ crates/kernel/src/arch/aarch64/syscall.rs
@@ -29,12 +29,12 @@ pub fn handle_syscall(frame: *mut u8) {
         let result: i64 = match nr {
             220 => {
                 CURRENT_REGS_PTR = regs;
-                crate::syscall::generic_vfork::<super::Aarch64>()
+                crate::syscall::generic_vfork::<super::Aarch64>() as i64
             }
-            221 => { crate::syscall::generic_exec::<super::Aarch64>(a0, a1); 0 }
+            221 => { crate::syscall::generic_exec::<super::Aarch64>(a0 as usize, a1 as usize); 0 }
             _ => {
                 let sc = translate_aarch64(nr);
-                crate::syscall::dispatch(sc, a0, a1, a2, a3, a4)
+                crate::syscall::dispatch(sc, a0 as usize, a1 as usize, a2 as usize, a3 as usize, a4 as usize) as i64
             }
         };
